% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_prefix}
\alias{add_prefix}
\title{Add a word prefix}
\usage{
add_prefix(texts, prefix)
}
\arguments{
\item{texts}{a \link{character} containing the original text}

\item{prefix}{a unit \link{character} containing the prefix to add (length == 1)}
}
\description{
Add a custom prefix to each word of a a line.
Number of spaces are normalized.
}
\examples{
add_prefix(c("this is a test", "this is another    test"), "#")
}
