% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{average_vectors}
\alias{average_vectors}
\title{Average vectors}
\usage{
average_vectors(keys, mat)
}
\arguments{
\item{keys}{\link{list} containing ids of embeddings in the matrix. Each slot of the \link{list} is related to a sequence.}

\item{mat}{embedding \link{matrix}}
}
\description{
Efficient implementation of a function to average vectors from a matrix.
}
\examples{
if (interactive()){
# This example should be run with a higher quality model
# than the one embedded in fastrtext
library(projector)
library(fastrtext)

model_test_path <- system.file("extdata",
                               "model_unsupervised_test.bin",
                               package = "fastrtext")
model <- load_model(model_test_path)
word_embeddings <- get_word_vectors(model,
                                    words = head(get_dictionary(model), 2e5))

average_vectors(strsplit(x = "this function average vector", split = " "), word_embeddings)
}
}
