% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{build_annoy_model}
\alias{build_annoy_model}
\title{Build Annoy model}
\usage{
build_annoy_model(vectors, number_trees)
}
\arguments{
\item{vectors}{\link{matrix} where each row is an observation}

\item{number_trees}{\link{integer} counting the number of trees to grow in Annoy (for neighboor search). More gives better results but is slower to compute.}
}
\description{
Annoy is used to retrieve the most similar vectors to a pivot one.
This function builds the Annoy model.
}
