% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{center_coordinates}
\alias{center_coordinates}
\title{Center coordinates around the pivot vector}
\usage{
center_coordinates(coordinates)
}
\arguments{
\item{coordinates}{\link{data.frame} containing coordinates to center. First position is the pivot.}
}
\description{
Center coordinates around the pivot vector
}
\keyword{internal}
