% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert from degrees to radians}
\usage{
deg2rad(x)
}
\arguments{
\item{x}{(numeric) Angle in degrees.}
}
\value{
\code{x} in radians.
}
\description{
Convert from degrees to radians
}
\examples{
\dontrun{
projecter::deg2rad(90)
}
}
