% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_from_metadata.R
\name{filter_from_metadata}
\alias{filter_from_metadata}
\title{Title}
\usage{
filter_from_metadata(
  .data,
  Y = NULL,
  metadata,
  filter_column,
  filter_value,
  sample_col = "sample"
)
}
\arguments{
\item{.data}{data.frame}

\item{Y}{Y}

\item{metadata}{metadata}

\item{filter_column}{column in metadata}

\item{filter_value}{value in filter_column}
}
\value{
a filter list of dataframe
}
\description{
Title
}
\examples{
data(iris)
.data <- iris[1:4]
rownames(.data) <- paste0("sample_", 1:nrow(.data))
Y = iris$Species \%>\% as.factor() \%>\% as.numeric()
metadata <- data.frame(sample =  paste0("sample_", 1:nrow(.data)), species = iris$Species)
filter_column <- "species"
filter_value = "versicolor"
filter_from_metadata(.data = .data, metadata = metadata,
filter_column = "species", filter_value = "versicolor")

filter_from_metadata(.data = .data, Y = Y, metadata = metadata,
filter_column = "species", filter_value = "versicolor")
}
