% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Compute 2D coordinates of vectors}
\usage{
get_coordinates(vectors, projection_type, max_iter = 500, perplexity = 30,
  verbose = FALSE)
}
\arguments{
\item{vectors}{\link{matrix} containing the \code{n} closest neighboor}

\item{projection_type}{\link{character} defining the algorithm to use to compute the coordinates. (\code{tsne} or \code{pca})}

\item{max_iter}{maximum number of epoch (for T-SNE learning)}

\item{perplexity}{how to balance attention between local and global aspects of data}

\item{verbose}{print debug information (for T-SNE learning)}
}
\description{
Transform original vectors in 2D coordinates applying \code{PCA} or \code{T-SNE}.
}
