% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{get_coordinates_pca}
\alias{get_coordinates_pca}
\title{Compute 2D coordinates using PCA}
\usage{
get_coordinates_pca(vectors)
}
\arguments{
\item{vectors}{\link{matrix} containing the \code{n} closest neighboor}
}
\description{
Rapid to compute but not always very meaningful.
}
