% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_coordinates_tsne}
\alias{get_coordinates_tsne}
\title{Compute 2D coordinates using T-SNE}
\usage{
get_coordinates_tsne(vectors, max_iter, verbose)
}
\arguments{
\item{vectors}{\link{matrix} containing the \code{n} closest neighboor}

\item{max_iter}{maximum number of epoch (for T-SNE learning)}

\item{verbose}{print debug information}
}
\description{
Better results but slow.
}
\keyword{internal}
