% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{get_ellipse}
\alias{get_ellipse}
\title{Title}
\usage{
get_ellipse(projectionObj, group = NULL, ellipse.level = 0.95)
}
\arguments{
\item{x}{projectionObj}
}
\value{
a tibble
}
\description{
Title
}
\examples{
data(iris)
.data <- iris[1:4]
rownames(.data) <- paste0("sample_", 1:nrow(.data))
.metadata <- data.frame(sample =  paste0("sample_", 1:nrow(.data)), species = iris$Species)
Y = iris$Species \%>\% as.factor() \%>\% as.numeric()
res_pca <- projection(.data, .metadata = .metadata, method = "pca")

res_ellipse <- get_ellipse(res_pca, group = "species")
res_ellipse <- get_ellipse(res_pca)

res_ellipse \%>\% dplyr::select(c(sample, group, inside_distance)) \%>\% unnest(cols = c(sample, inside_distance))
}
