% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Retrieve the most closest vector representation}
\usage{
get_neighbors(word, dict, annoy_model, n, search_k)
}
\arguments{
\item{word}{\link{character} containing the pivot word}

\item{dict}{\link{character} containing all possible texts}

\item{annoy_model}{\link{RcppAnnoy} model}

\item{n}{number of elements to retrieve}

\item{search_k}{number of nodes to search in (Annoy parameter). Higher is better and slower.}
}
\description{
Use Annoy to rapidly retrieve the \code{n} most closest representation of a text.
}
\keyword{internal}
