% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_neighbors_from_free_text}
\alias{get_neighbors_from_free_text}
\title{Retrieve the most closest vector representation of a free text}
\usage{
get_neighbors_from_free_text(text, annoy_model, n, search_k, projector_instance)
}
\arguments{
\item{text}{\link{character} containing the text searched.}

\item{annoy_model}{\link{RcppAnnoy} model}

\item{n}{top n elements to retrieve}

\item{search_k}{number of nodes to search in (Annoy parameter). Higher is better (and slower).}

\item{projector_instance}{an instance of projector tool generated by \link{get_projector_instance}}
}
\description{
Use Annoy to rapidly retrieve the \code{n} most closest representation of a text.
}
\keyword{internal}
