% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_annoy.R
\name{get_neighbors_from_text}
\alias{get_neighbors_from_text}
\title{Retrieve the most closest vector representation of an indexed text}
\usage{
get_neighbors_from_text(text, annoy_model, n, search_k)
}
\arguments{
\item{text}{\link{character} containing the pivot text}

\item{annoy_model}{\link{RcppAnnoy} model}

\item{n}{number of elements to retrieve}

\item{search_k}{number of nodes to search in (\link{RcppAnnoy} parameter). Higher is better and slower.}
}
\description{
Use \link{RcppAnnoy} to rapidly retrieve the \code{n} most closest representation of a text.
The text has to be already indexed in \link{RcppAnnoy}.
If it is not the case, use \link{get_neighbors_from_free_text}.
}
