% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_annoy.R
\name{get_vector_from_text}
\alias{get_vector_from_text}
\title{Retrieve vector representation  of an indexed text}
\usage{
get_vector_from_text(text, annoy_model)
}
\arguments{
\item{text}{\link{character} containing the text}

\item{annoy_model}{\link{RcppAnnoy} model}
}
\description{
Retrieve from \link{RcppAnnoy} the vector related to an indexed text.
}
\examples{
if (interactive()){
library(projector)
library(fastrtext)

model_test_path <- system.file("extdata",
                               "model_unsupervised_test.bin",
                               package = "fastrtext")
model <- load_model(model_test_path)
word_embeddings <- get_word_vectors(model,
                                    words = head(get_dictionary(model), 2e5))

annoy_model <- get_annoy_model(word_embeddings, 5)

print(get_vector_from_text("the", annoy_model))
}
}
