% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_annoy.R
\name{get_word_position}
\alias{get_word_position}
\title{Get word position}
\usage{
get_word_position(word, annoy_model)
}
\arguments{
\item{word}{\link{character} words to get the position from}

\item{annoy_model}{\link{RcppAnnoy} model}
}
\description{
Get word position
}
\keyword{internal}
