% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{\%where\%}
\alias{\%where\%}
\title{Return positions of items in a \link{character} efficiently}
\usage{
items \%where\% collection
}
\arguments{
\item{items}{\link{character} with elements to find position}

\item{collection}{\link{character} where to search}
}
\value{
\link{integer} with position. May be shorter than \code{items} if one or more elements are not found in \code{collection}
}
\description{
First use index elements, then rapidly retrieve the positions.
}
