% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{interactive_embedding_exploration}
\alias{interactive_embedding_exploration}
\title{Interactive exploration of embeddings}
\usage{
interactive_embedding_exploration(annoy_model)
}
\arguments{
\item{annoy_model}{\link{RcppAnnoy} model generated with \link{get_annoy_model}}
}
\description{
Shiny application to interactively play with embeddings.
User provides a pivot word and the n most similar word
are projected on a scatter plot.
}
