% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.projection}
\alias{plot.projection}
\title{plot_projectionObj}
\usage{
\method{plot}{projection}(
  x,
  color = NULL,
  shape = NULL,
  size = 3,
  show_names = TRUE,
  title = NULL,
  legend = TRUE,
  legend.position = "right",
  legend.box = "vertical",
  graph = TRUE
)
}
\arguments{
\item{x}{projectionObj}

\item{color}{The name of the column in \code{projectionObj$coord} to use to color the
points in the plot. If \code{NULL}, won't add color. Default: \code{NULL}.}

\item{shape}{The name of the column in \code{res_pca$coord} to use to define
the shape of the points in the PCA. If \code{NULL}, won't add change the shape.
Default: \code{NULL}.}

\item{size}{The size of the points. Default: 3.}

\item{show_names}{Try to add the labels of each point? Will only add the
labels if the points are not too closely clustered together in the PCA.
Default: \code{TRUE}.}

\item{title}{Add a title to the graph? If \code{NULL}, no title is added to
the graph. Otherwise this param must be a \code{character}. Default:
\code{NULL}.}

\item{legend}{description}

\item{legend.position}{Value for the \code{legend.position} param from
ggplot2. Default: "right".}

\item{legend.box}{description}

\item{graph}{Produce the graph. \code{TRUE} or \code{FALSE}. Default:
\code{TRUE}.}
}
\description{
a short description
}
