% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_projection}
\alias{plot_projection}
\title{Plot projection}
\usage{
plot_projection(projection, ...)
}
\arguments{
\item{projection}{Projection function.}

\item{...}{Additional arguments passed to \code{projection}}
}
\value{
ggplot object
}
\description{
Maps a grid of angles using a particular projection.
}
\examples{
\dontrun{
projector::plot_projection(projection = projector::aitorff)
}
}
