% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{plot_text}
\alias{plot_text}
\title{Plot vectors on a 2D plan}
\usage{
plot_text(coordinates, min_cluster_size = 5)
}
\arguments{
\item{coordinates}{\link{data.frame} containing 2D coordinates of texts.}

\item{min_cluster_size}{\link{integer} corresponding to the minimum size of a vector (for the colors).}
}
\description{
Plot the text on an interactive scatter plot.
}
