% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_projection_produce_df.R
\name{produce_pca_df}
\alias{produce_pca_df}
\title{produce_pca_df}
\usage{
produce_pca_df(.data, ...)
}
\arguments{
\item{.data}{a matrix}
}
\value{
a list
}
\description{
This function produce a dataframe based on a 2 components PCA sample projection.
}
\examples{
data('iris')
X <- iris
projection.obj <- produce_pca_df(X)

}
