% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_projection_produce_df.R
\name{produce_umap_df}
\alias{produce_umap_df}
\title{produce_umap_df}
\usage{
produce_umap_df(.data, n_components = 2, random_state = 1, ...)
}
\arguments{
\item{.data}{a matrix}
}
\value{
a list
}
\description{
This function produce a dataframe based on a 2 components umap sample projection.
}
\examples{
data('iris')
X <- iris
projection.obj <- produce_umap(X)
}
