% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_table.R
\name{project_table}
\alias{project_table}
\title{Project a bimodal edge list into a unimodal edge list}
\usage{
project_table(df, joining_col, projected_col)
}
\arguments{
\item{df}{A data.frame with the oribinal bimodal edge list}

\item{joining_col}{The quoted name of the column with tth}
}
\description{
Project a bimodal edge list into a unimodal edge list
}
\examples{
bipartite_list <- data.frame(
  club = c("Karate", "Karate", "Book", "Chess", "Book", "Chess"),
  student = c(1, 2, 1, 3, 2, 4),
  stringsAsFactors = FALSE)

project_table(bipartite_list, joining_col = "club", projected_col = "student")
#>   from to weight
#> 1    1  2      2
#> 2    3  4      1
}

