% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{projection}
\alias{projection}
\title{projection}
\usage{
projection(
  .data,
  method = "pca",
  .metadata = NULL,
  .metadata_join_col = NULL,
  ...
)
}
\arguments{
\item{.data}{a numeric data.frame}

\item{method}{character, available methods: pca}

\item{.metadata}{(optional) a data.frame with sample metadata}

\item{.metadata_join_col}{(optional) if .metadata is given, .metadata_join_col must be a column of .metadata. Otherwise, .metadata_join_col <- "sample".}

\item{...}{other argument to pass to projection method}
}
\description{
This function can be used to call different projection methods on a data.frame
using the same interface.
}
