% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projections}
\alias{projections}
\alias{aitorff}
\alias{equirectangular}
\alias{winkel_bopc}
\title{Map projections}
\usage{
aitorff(lambda, phi, R = 6378, radians = FALSE)

equirectangular(
  lambda,
  phi,
  R = 6378,
  lambda_0 = 0,
  phi_0 = 0,
  standard_parallel = 0,
  radians = FALSE
)

winkel_bopc(
  lambda,
  phi,
  R = 6378,
  lambda_0 = 0,
  phi_0 = 0,
  standard_parallel = 0,
  radians = FALSE
)
}
\arguments{
\item{lambda}{(numeric) Longitude.}

\item{phi}{(numeric) Latitude.}

\item{R}{(numeric) Radius of the globe.}

\item{radians}{(Boolean) Are the angles specified in radians?}

\item{lambda_0}{(numeric) Central parallel.}

\item{phi_0}{(numeric) Central meridian}

\item{standard_parallel}{(numeric) Standard parallel where the scale of the
projection is true.}
}
\value{
\code{x} and \code{y} coordinates of equirectangular projection.
}
\description{
Map projections
}
\examples{
\dontrun{
}
}
