% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Projector}
\alias{projector}
\alias{projector-package}
\alias{Projector-package}
\title{projector: Project Dense Vectors Representation of Texts on a 2D Plan}
\description{
Display dense vector representation of texts on a 2D plan to better
understand embeddings by observing the neighbors of a selected text.
It also includes an interactive application to change dynamically the pivot text.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pommedeterresautee/projector}
  \item Report bugs at \url{https://github.com/pommedeterresautee/projector/issues}
}

}
\author{
\strong{Maintainer}: Michaël Benesty \email{michael@benesty.fr} [copyright holder]

}
