% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projector_setup.R
\name{projector_setup}
\alias{projector_setup}
\title{Set up a folder structure for a projector project}
\usage{
projector_setup(sub_projects = 1)
}
\arguments{
\item{sub_projects}{number of desired subprojects (or flat if 0). Defaults to 1.
Call this function only within existing root folder.}
}
\description{
Populates a folder with default folder structure. Usually called from within projector_create()
}

