% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Rcpp_projector-class}
\alias{Rcpp_projector-class}
\title{Rcpp_projector class}
\description{
Efficient implementation of a function to average embeddings of words from a sentence.

Constructor takes a \link{matrix} (\code{embeddings}) where each row is a word embedding (each row should have a name, the name being the word related to the embedding) and a \link{logical} (\code{na_if_unknwown_word_p}) set to \link{TRUE} if it is expected to fulfill a row with \link{NA} when at the minimum one word of the sentence is unknown, and set to \link{FALSE} where it is expected to only average vectors related to known words. If no word are known, the row will always be fulfilled by \link{NA}, whatever the parameters are.

Words are defined as letters between separated by one or more spaces in a sentence.
}
\section{Slots}{

\describe{
\item{\code{average_vectors}}{provide \link{character} (\code{texts}) and return a \link{matrix} of sentence embeddings.}

\item{\code{set_unknown_word}}{change value of parameter \code{na_if_unknwown_word_p} (see above).}
}}

