% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{retrieve_neighbors}
\alias{retrieve_neighbors}
\title{Retrieve a list of neighbor vectors}
\usage{
retrieve_neighbors(text, projection_type, annoy_model, n,
  search_k = max(10000, 10 * n), ...)
}
\arguments{
\item{text}{\link{character} containing the text related to the pivot vector}

\item{projection_type}{\link{character} defining the algorithm to use to compute the coordinates}

\item{annoy_model}{\link{RcppAnnoy} model}

\item{n}{number of elements to retrieve}

\item{search_k}{number of nodes to search in (Annoy parameter). Higher is better and slower.}

\item{...}{additional parameters used in \link{get_coordinates}}
}
\description{
Use \link{RcppAnnoy} to rapidly retrieve a list of vector neighbors
}
