% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{save_annoy_model}
\alias{save_annoy_model}
\title{Save \link{RcppAnnoy} model}
\usage{
save_annoy_model(annoy_model, path_annoy, path_dictionary)
}
\arguments{
\item{annoy_model}{\link{RcppAnnoy} model}

\item{path_annoy}{path for the \link{RcppAnnoy} model}

\item{path_dictionary}{path for the dictionary (\link{character} containing texts)}
}
\description{
Save the content of the model in two files:
\itemize{
\item the \link{RcppAnnoy} model
\item the dictionary (\link{character} containing texts)
}
}
