% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_text.R
\name{shiny_text}
\alias{shiny_text}
\title{Interactive approximative search of embeddings (text)}
\usage{
shiny_text(annoy_model, word_embeddings_mat, normalize_text)
}
\arguments{
\item{annoy_model}{\link{RcppAnnoy} model generated with \link{get_annoy_model}}

\item{word_embeddings_mat}{\link{matrix} containing embeddings of each word (used to generate the embedding of the query)}

\item{normalize_text}{\link{function} to clean the query before being transformed in embedding. Take a \link{character} as parameter and return a \link{character}.}
}
\description{
Shiny application to display closest embeddings to a query
User provides a pivot word and the n most similar word
are displayed.
}
\details{
Word embeddings are retrieved from the provided \link{matrix} and averaged to get the embedding of the query.
}
\examples{
if(interactive()){
# This example should be run with a higher quality model
# than the one embedded in fastrtext

library(projector)
library(fastrtext)

model_test_path <- system.file("extdata",
                               "model_unsupervised_test.bin",
                               package = "fastrtext")

model <- load_model(model_test_path)
word_embeddings_mat <- get_word_vectors(model, words = head(get_dictionary(model), 2e5))
annoy_model <- get_annoy_model(word_embeddings_mat, 5)

shiny_text(annoy_model = annoy_model,
           word_embeddings_mat = word_embeddings_mat,
           normalize_text = tolower)
}
}
