% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_project.R
\name{start_project}
\alias{start_project}
\title{Initialize Research Project Structure}
\usage{
start_project(
  path = ".",
  project_name = NULL,
  study_name = "study-1",
  overwrite = FALSE,
  literate = TRUE,
  git_init = FALSE,
  open_session = FALSE
)
}
\arguments{
\item{path}{Directory where the project should be created. If it doesn't exist, it will be created. Use "." to work in current directory.}

\item{project_name}{Name of the overall research project. If NULL, uses the basename of the path.}

\item{study_name}{Name of the study folder within studies/ (default: "study-1")}

\item{overwrite}{Whether to overwrite existing files}

\item{literate}{If TRUE, creates .qmd files for literate programming. If FALSE, creates .R scripts with a master source file}

\item{git_init}{If TRUE, initializes a git repository with initial commit}

\item{open_session}{If TRUE, opens .Rproj of newly created project}
}
\description{
Creates a standardized research project structure with optional literate programming and git initialization
}
\examples{
\dontrun{
# Create new project in new directory
start_project("~/research/my-awesome-study")

# Create project in current directory with custom name
start_project(".", project_name = "dissertation-study", study_name = "experiment-1")

# Create with R scripts instead of .qmd files
start_project("my-project", study_name = "pilot", literate = FALSE, git_init = TRUE)
}
}
