% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Compute 2D coordinates of vectors}
\usage{
get_coordinates(vectors, projection_type)
}
\arguments{
\item{vectors}{\link{matrix} containing the \code{n} closest neighboor}

\item{projection_type}{\link{character} defining the algorithm to use to compute the coordinates. \code{tsne} or \code{pca}}
}
\description{
Plot function
}
