% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Compute 2D coordinates of vectors}
\usage{
get_coordinates(vectors, projection_type, ...)
}
\arguments{
\item{vectors}{\link{matrix} containing the \code{n} closest neighbor}

\item{projection_type}{\link{character} defining the algorithm to use to compute the coordinates. (\code{tsne} or \code{pca})}

\item{...}{parameters passed to projection algorithm (\code{max_iter}, \code{perplexity}, \code{verbose}, \code{transformations})}
}
\description{
Transform original vectors in 2D coordinates applying either:
\itemize{
\item \href{https://en.wikipedia.org/wiki/Principal_component_analysis}{PCA}
\item \href{https://distill.pub/2016/misread-tsne/}{T-SNE}.
}
}
\keyword{internal}
