% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_coordinates_pca}
\alias{get_coordinates_pca}
\title{Compute 2D coordinates using PCA}
\usage{
get_coordinates_pca(vectors, transformations = c("center", "scaled"), ...)
}
\arguments{
\item{vectors}{\link{matrix} containing the \code{n} closest neighbor}

\item{transformations}{transformations applied to vectors before computing PCA}
}
\description{
Rapid to compute but not always very meaningful.
Data are centered and scaled.
}
\keyword{internal}
