% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_neighbors_from_free_text}
\alias{get_neighbors_from_free_text}
\title{Retrieve the most closest vector representation of a free unknown text}
\usage{
get_neighbors_from_free_text(text, annoy_model, word_embeddings_mat, n,
  search_k)
}
\arguments{
\item{text}{\link{character} containing the pivot text. Can be unknown text.}

\item{annoy_model}{\link{RcppAnnoy} model}

\item{word_embeddings_mat}{a \link{matrix} containing word embeddings.}

\item{n}{number of elements to retrieve}

\item{search_k}{number of nodes to search in (Annoy parameter). Higher is better and slower.}
}
\description{
Use Annoy to rapidly retrieve the \code{n} most closest representation of a text.
}
\keyword{internal}
