% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_neighbors_from_text}
\alias{get_neighbors_from_text}
\title{Retrieve the most closest vector representation of a text}
\usage{
get_neighbors_from_text(text, annoy_model, n, search_k)
}
\arguments{
\item{text}{\link{character} containing the pivot text}

\item{annoy_model}{\link{RcppAnnoy} model}

\item{n}{number of elements to retrieve}

\item{search_k}{number of nodes to search in (Annoy parameter). Higher is better and slower.}
}
\description{
Use Annoy to rapidly retrieve the \code{n} most closest representation of a text.
}
\keyword{internal}
