% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_vector_from_text}
\alias{get_vector_from_text}
\title{Retrieve vector representation of a text}
\usage{
get_vector_from_text(text, annoy_model)
}
\arguments{
\item{text}{\link{character} containing the text}

\item{annoy_model}{\link{RcppAnnoy} model}
}
\description{
Use dictionary to retrieve the vector stored in annoy.
}
