% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{retrieve_neighbors}
\alias{retrieve_neighbors}
\title{Retrieve a list of neighbor vectors}
\usage{
retrieve_neighbors(text, embeddings, projection_type, model, n,
  search_k = min(max(10000, 10 * n), nrow(embeddings)))
}
\arguments{
\item{text}{\link{character} containing the text related to the pivot vector}

\item{embeddings}{\link{matrix} containing all possible vectors}

\item{projection_type}{\link{character} defining the algorithm to use to compute the coordinates}

\item{model}{\link{RcppAnnoy} model}

\item{n}{number of elements to retrieve}

\item{search_k}{number of nodes to search in (Annoy parameter). Higher is better and slower.}
}
\description{
Use \link{RcppAnnoy}
}
